fix: test path directly
